+++ src/Control/Ether/TH.hs
@@ -1,11 +1,9 @@
 {-# LANGUAGE TemplateHaskell #-}
 
--- | Template haskell utilities.
+-- | Template Haskell utilities.
 
 module Control.Ether.TH
     ( ethereal
-    , fmapN
-    , deepN
     ) where
 
 import qualified Language.Haskell.TH as TH
@@ -40,19 +38,3 @@ ethereal strTagName strTagProxyName = do
     tagDecl <- emptyDataDecl tagName
     (tagProxySig, tagProxyVal) <- proxySimple tagProxyName tag
     return [tagDecl, tagProxySig, tagProxyVal]
-
--- |
--- Compose 'fmap' @n@ times.
---
--- @$('fmapN' 5) = fmap.fmap.fmap.fmap.fmap@
-fmapN :: Int -> TH.ExpQ
-fmapN 0 = [|id|]
-fmapN n = TH.infixApp [|fmap|] [|(.)|] (fmapN (n - 1))
-
--- |
--- 'fmap' a function @n@ levels deep.
---
--- @$('deepN' 3 [| f |]) = $('fmapN' 3) f ($('fmapN' 2) f ($('fmapN' 1) f id))@
-deepN :: Int -> TH.ExpQ -> TH.ExpQ
-deepN 0 _ = [|id|]
-deepN n f = [| $(fmapN n) $f $(deepN (n-1) f) |]