+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-0.1.1.0
+0.2.0.0
 -------
 
 * Convenience modules `Control.Monad.Ether` and `Control.Monad.Ether.Implicit`.
+* Remove `fmapN` and `deepN`.
+* Remove `Control.Monad.Ether.Implicit.Except.TH`.
+* Add `handle` and `handleT`.
 
 
 0.1.0.1